@@ -4,12 +4,8 @@ namespace Prescreen\ApiResourceBundle\Application\Configuration\FieldOptions;
 
 class BoolField extends FieldOptions
 {
-    const TYPE = 'bool';
+    const string TYPE = 'bool';
 
-    /**
-     * @param bool $required
-     * @param bool $default
-     */
     public function __construct(bool $required = false, bool $default = false)
     {
         parent::__construct(self::TYPE, $required, $default);
