@@ -12,19 +12,12 @@ use Prescreen\ApiResourceBundle\Exception\ValueNotAllowedException;
 class StringValidator extends ApiValidator
 {
     /**
-     * @param string $fieldName
-     * @param $value
-     * @param FieldOptions $fieldOptions
-     * @param $oldValue
-     *
      * @throws FieldTypeException
      * @throws RequiredFieldMissingException
      * @throws PermissionDeniedException
      * @throws ValueNotAllowedException
-     *
-     * @return mixed
      */
-    public function validate(string $fieldName, $value, FieldOptions $fieldOptions, $oldValue = null)
+    public function validate(string $fieldName, $value, FieldOptions $fieldOptions, $oldValue = null): ?string
     {
         parent::validate($fieldName, $value, $fieldOptions, $oldValue);
 
