@@ -2,6 +2,6 @@
 
 namespace Prescreen\ApiResourceBundle\Exception;
 
-class FieldTypeException extends InternalApiValidatorException
+class FieldTypeException extends ApiValidatorException
 {
 }
