@@ -2,6 +2,6 @@
 
 namespace Prescreen\ApiResourceBundle\Exception;
 
-class ValueNotAllowedException extends InternalApiValidatorException
+class ValueNotAllowedException extends ApiValidatorException
 {
 }
